refactor: move Documentation Table render to separate module DocumentationTable.jsx
