+++ src/scripts/modules/storage-explorer/react/pages/Documentation/Documentation.jsx
@@ -1,22 +1,23 @@
 import React from 'react';
 import createReactClass from 'create-react-class';
 import createStoreMixin from '../../../../../react/mixins/createStoreMixin';
-import { Table, SplitButton, MenuItem } from 'react-bootstrap';
+import { SplitButton, MenuItem } from 'react-bootstrap';
 import NavButtons from '../../components/NavButtons';
 import { Navigation } from 'react-router';
 import ApplicationActionCreators from '../../../../../actions/ApplicationActionCreators';
+import DocumentationTable from './DocumentationTable';
 
 import BucketsStore from '../../../../components/stores/StorageBucketsStore';
 import TablesStore from '../../../../components/stores/StorageTablesStore';
 import FilesStore from '../../../../components/stores/StorageFilesStore';
 import ApplicationStore from '../../../../../stores/ApplicationStore';
 import DocumentationLocalStore from '../../../DocumentationLocalStore';
-import Markdown from '../../../../../react/common/Markdown';
+
 import { SearchBar, Loader } from '@keboola/indigo-ui';
 
-import {createDocumentationTree, rowTypes, buildDocumentationToMarkdown} from '../../../DocumentationUtils';
+import {createDocumentationTree, buildDocumentationToMarkdown} from '../../../DocumentationUtils';
 
-import { toggleDocumentationRow, updateDocumentationSearchQuery, uploadFile, loadLastDocumentationSnapshot } from '../../../Actions';
+import {toggleDocumentationRow, updateDocumentationSearchQuery, uploadFile, loadLastDocumentationSnapshot } from '../../../Actions';
 
 
 
@@ -84,9 +85,12 @@ export default createReactClass({
                 </SplitButton>
               }
             />
-            <Table striped responsive>
-              <tbody>{this.renderDocumentationTree()}</tbody>
-            </Table>
+            <DocumentationTable
+              documentationTree={this.state.documentationTree}
+              openedRows={this.state.openedRows}
+              searchQuery={this.state.searchQuery}
+              toggleDocumentationRow={toggleDocumentationRow}
+            />
           </div>
         </div>
       </div>
@@ -115,70 +119,5 @@ export default createReactClass({
       ApplicationActionCreators.sendNotification({ message: 'Storage documentation snapshot created' });
       return loadLastDocumentationSnapshot();
     });
-  },
-
-  renderDocumentationTree() {
-    return this.state.documentationTree.reduce((bucketsMemo, bucket) => {
-      const bucketId = bucket.get('id');
-      bucketsMemo.push(
-        this.renderOneTableRow(bucketId, bucketId, bucket.get('bucketDescription'), rowTypes.BUCKET_ROW)
-      );
-      if (!this.state.openedRows.get(rowTypes.BUCKET_ROW + bucketId) && !this.state.searchQuery) {
-        return bucketsMemo;
-      }
-      const bucketTablesRows = bucket.get('bucketTables').reduce((tablesMemo, table) => {
-        const tableId = table.get('id');
-        tablesMemo.push(
-          this.renderOneTableRow(tableId, table.get('name'), table.get('tableDescription'), rowTypes.TABLE_ROW)
-        );
-        if (!this.state.openedRows.get(rowTypes.TABLE_ROW + tableId) && !this.state.searchQuery) {
-          return tablesMemo;
-        }
-        const columnsRows = table.get('columnsDescriptions').reduce((columnsMemo, description, column) => {
-          // const description = table.getIn(['columnsDescriptions', column]);
-          columnsMemo.push(this.renderOneTableRow(tableId + column, column, description, rowTypes.COLUMN_ROW));
-          return columnsMemo;
-        }, []);
-        return tablesMemo.concat(columnsRows);
-      }, []);
-      return bucketsMemo.concat(bucketTablesRows);
-    }, []);
-  },
-
-  renderOneTableRow(id, name, description, rowType) {
-    const pointerClass = !this.state.searchQuery ? 'kbc-cursor-pointer ' : '';
-    let divClassName = pointerClass + 'bucket-row';
-
-    let rowTypeClassName = 'fa fa-folder';
-    if (rowType === rowTypes.TABLE_ROW) {
-      divClassName = pointerClass + 'table-row';
-      rowTypeClassName = 'fa fa-table';
-    }
-    if (rowType === rowTypes.COLUMN_ROW) {
-      divClassName = 'column-row';
-      rowTypeClassName = 'fa fa-columns';
-    }
-    const isOpened = this.state.openedRows.get(rowType + id);
-    const caretClass = isOpened ? 'fa fa-caret-down' : 'fa fa-caret-right';
-    const caret = !this.state.searchQuery ? <i className={caretClass} /> : null;
-
-    return (
-      <tr key={id}>
-        <td className="text-nowrap">
-          <div className={divClassName} onClick={() => toggleDocumentationRow(rowType + id, !isOpened)}>
-            {rowType !== rowTypes.COLUMN_ROW && caret}
-            {' '}
-            <i className={rowTypeClassName} />
-            {' '}
-            {name}
-          </div>
-        </td>
-        <td className="kbc-break-all kbc-break-word">
-          {description ? <Markdown source={description} collapsible={true} /> : 'N/A'}
-        </td>
-      </tr>
-    );
-  },
-
-
+  }
 });
